+++ src/samples/sample-piral/src/components/index.ts
@@ -2,3 +2,4 @@ export * from './MenuToggle';
 export * from './User';
 export * from './LanguageIcon';
 export * from './LanguagePicker';
+export * from './Layout';
