+++ src/samples/sample-piral/src/styles/_layout.scss
@@ -57,8 +57,13 @@ body.pi-modal-open {
   }
 }
 
+.text-right {
+  text-align: right;
+}
+
 .language-picker {
   position: relative;
+  margin: 4px 8px 0 0;
 
   .current {
     cursor: pointer;
@@ -70,14 +75,13 @@ body.pi-modal-open {
     border: none;
     margin: 0;
     padding: 0;
-    width: 200px;
     overflow: hidden;
     border-radius: 3px 0 3px 3px;
     background-clip: padding-box;
-    margin-right: 5px;
+    margin-bottom: 5px;
     position: absolute;
-    bottom: 100%;
-    left: 0;
+    top: 100%;
+    right: 0;
 
     li {
       cursor: pointer;
@@ -86,7 +90,13 @@ body.pi-modal-open {
       margin: 0;
       outline: 0;
       color: #7d8086;
+      display: flex;
+      flex-direction: row;
+      align-items: center;
 
+      span {
+        padding-left: 8px;
+      }
 
       &:hover {
         background: rgba(235, 235, 235, 0.9);
